feat!: Implement middleware system for modifying requests/responses
Middleware can now be added to an HttpClient. Each Middleware is simply
an async function that receives the request-to-be-sent, and can
optionally modify it before passing it further down the chain by calling
the `next()` parameter. The final middleware's call to `next()` sends
the request via the NetworkAdapter. Then, the response is passed back
down the stack and is available in each middleware by awaiting `next()`.
Again, each middleware can choose to modify the response.

Cookie sending/storing is implemented as a middleware now, decoupling
HttpClient from dealing with cookies entirely.
Another use case for middleware would be to set headers, for example.

BREAKING CHANGE: Only StatefulHttpClient offers setCookie() now.