fix!: Use `Headers.prototype.getSetCookie` to avoid comma-splitting
Since Node.js now has `getSetCookie` available, we can use it instead of
splitting the header on comma, which was very error-prone.

Ref: https://github.com/nodejs/undici/pull/1915

The linked PR was released as Undici v5.19.0, which is first available
in Node.js 18.14.1. Since Node.js 18.14.2 has an even more up-tp-date
version of Undici (5.20.0), I figured it'd be sensible to raise our
requirement to that.

BREAKING CHANGE: Engine version requirement raised to Node `>=18.14.2`.