feat: Parse "Expires" cookie attribute
This adds a dependency on "set-cookie-parser", which does the heavy
lifting of cookie parsing. I thought about implementing this myself, but
it's simply not worth it. For now, our memory cookie store makes use of
the "expires" attribute so that cookies have a maximum lifetime, and can
also be deleted by the server by setting "expires" to a past date.